Material and themes added
