+++ app/src/main/java/android/game/simplecalc/controller/MainController.java
@@ -1,0 +1,45 @@
+package android.game.simplecalc.controller;
+
+import android.widget.TextView;
+
+public interface MainController {
+
+    void calcActionAC(TextView calcField);
+
+    void calcActionClear(TextView calcField);
+
+    void calcActionPercent(TextView calcField);
+
+    void calcActionShare(TextView calcField);
+
+    void calcActionMultiply(TextView calcField);
+
+    void calcActionPlus(TextView calcField);
+
+    void calcActionMinus(TextView calcField);
+
+    void calcActionEquals(TextView calcField);
+
+    void calcActionPoint(TextView calcField);
+
+    void calcActionButton1(TextView calcField);
+
+    void calcActionButton2(TextView calcField);
+
+    void calcActionButton3(TextView calcField);
+
+    void calcActionButton4(TextView calcField);
+
+    void calcActionButton5(TextView calcField);
+
+    void calcActionButton6(TextView calcField);
+
+    void calcActionButton7(TextView calcField);
+
+    void calcActionButton8(TextView calcField);
+
+    void calcActionButton9(TextView calcField);
+
+    void calcActionButton0(TextView calcField);
+
+}
