+++ app/src/main/java/android/game/simplecalc/model/Counter.java
@@ -1,0 +1,7 @@
+package android.game.simplecalc.model;
+
+public interface Counter {
+
+    String countCalcField(String str);
+
+}
