+++ app/src/main/res/values/dimens.xml
@@ -1,12 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
     <dimen name="button_text">20sp</dimen>
-
-
     <dimen name="dimensionbetweenbuttons">6dp</dimen>
-
     <dimen name="dimensionbetweenbuttonsland">2dp</dimen>
-
-
 </resources>
