+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,24 @@
 <resources>
     <string name="app_name">SimpleCalc</string>
+    <string name="point">.</string>
+    <string name="_0">0</string>
+    <string name="_9">9</string>
+    <string name="_8">8</string>
+    <string name="_7">7</string>
+    <string name="_6">6</string>
+    <string name="_5">5</string>
+    <string name="_4">4</string>
+    <string name="_3">3</string>
+    <string name="_2">2</string>
+    <string name="_1">1</string>
+    <string name="_equals">=</string>
+    <string name="plus">+</string>
+    <string name="minus">—</string>
+    <string name="multiply">x</string>
+    <string name="share">/</string>
+    <string name="percent">%</string>
+    <string name="c">C</string>
+    <string name="ac">AC</string>
+    <string name="day">День</string>
+    <string name="night">Ночь</string>
 </resources>
